mend.bbclass: Ignore patched CVEs
Currently, the mend scan is launched without taking into account the
CVEs that are patched by the Yocto Project.
This leads to the report containing vulnerabilities which are not
actually present in the project.

Fix this by taking the patched CVEs and using the Mend API to set the
vulnerability alerts caused by them to be ignored.

Due to the fact that it is not feasible to add additional external
dependencies to the BitBake engine's internal Python, which is separate
from the native host Python, the function uses urllib and json to
internally build and put the HTTP requests.

Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>