mend: Move config outside of docker
It is not a great approach to have the config strictly tied to the
docker container, because this means having to release a new version of
the container each time one wants to modify it.

Fix this by adding a new variable to tell BitBake the path from which it
should take the config file.
This brings the additional advantage of being able to overwrite that
specific variable in a single package, to allow for custom package-based
configs instead of being forced to have the same global one for the
entire product.

A default `mend.wss.config` is still included in the files for this
meta-layer and is automatically accessed if another isn't specified.

Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>