mend.bbclass: Move from `e.data` to `d`
The documentation suggests using datastore (e.g. `d.getVar`) to obtain
the value of BitBake variables as opposed to `e.data.getVar`, which is
also much less common.
This is also what is being used in every other place within this
bbclass, with the exception of checking the required variables for the
unified agent.

Therefore, move the checking to this solution.

Link: https://docs.yoctoproject.org/bitbake/bitbake-user-manual/bitbake-user-manual-metadata.html
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>