mend.bbclass: Remove dependency from WS_PRODUCTTOKEN
To allow programmatically setting the name of the product at runtime, it
is mandatory to remove the dependency from its unique token.

To make this possible, add an additional check to create a new product
with the set WS_PRODUCTNAME if it doesn't yet exist.
After, simply query the API about the product token given the product
name.

Adjust README accordingly.

Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>